refactor!: remove cross-fetch
BREAKING CHANGE: This changes exports slightly, and requires Node.js
16.15 or higher / browsers newer than 2017 or so.
