+++ package.json
@@ -36,7 +36,6 @@
     "arraybuffer-loader": "^1.0.3",
     "base64-js": "^1.3.0",
     "buffer": "6.0.3",
-    "cross-fetch": "^4.1.0",
     "fastestsmallesttextencoderdecoder": "^1.0.7",
     "js-md5": "^0.7.3",
     "minilog": "^3.1.0"
