+++ src/scratchFetch.ts
@@ -1,6 +1,4 @@
-import * as crossFetch from 'cross-fetch';
-
-export const Headers = crossFetch.Headers;
+export const Headers = globalThis.Headers;
 
 /**
  * Metadata header names.
@@ -16,7 +14,7 @@ export enum RequestMetadata {
 /**
  * Metadata headers for requests.
  */
-const metadata = new crossFetch.Headers();
+const metadata = new Headers();
 
 /**
  * Check if there is any metadata to apply.
@@ -53,13 +51,13 @@ export const hasMetadata = (): boolean => {
 export const applyMetadata = (options?: globalThis.RequestInit): globalThis.RequestInit | undefined => {
     if (hasMetadata()) {
         const augmentedOptions = Object.assign({}, options);
-        augmentedOptions.headers = new crossFetch.Headers(metadata);
+        augmentedOptions.headers = new Headers(metadata);
         if (options && options.headers) {
             // the Fetch spec says options.headers could be:
             // "A Headers object, an object literal, or an array of two-item arrays to set request's headers."
             // turn it into a Headers object to be sure of how to interact with it
-            const overrideHeaders = options.headers instanceof crossFetch.Headers ?
-                options.headers : new crossFetch.Headers(options.headers);
+            const overrideHeaders = options.headers instanceof Headers ?
+                options.headers : new Headers(options.headers);
             for (const [name, value] of overrideHeaders.entries()) {
                 augmentedOptions.headers.set(name, value);
             }
@@ -79,7 +77,7 @@ export const applyMetadata = (options?: globalThis.RequestInit): globalThis.Requ
  */
 export const scratchFetch = (resource: RequestInfo | URL, options?: globalThis.RequestInit): Promise<Response> => {
     const augmentedOptions = applyMetadata(options);
-    return crossFetch.fetch(resource, augmentedOptions);
+    return fetch(resource, augmentedOptions);
 };
 
 /**
