+++ test/integration/download-known-assets.test.js
@@ -1,6 +1,9 @@
 const md5 = require('js-md5');
 
+const mockFetch = require('../fixtures/mockFetch.js');
 const ScratchStorage = require('../../src/index').ScratchStorage;
+
+jest.spyOn(global, 'fetch').mockImplementation(mockFetch.fetch);
 
 test('constructor', () => {
     const storage = new ScratchStorage();
@@ -68,7 +71,7 @@ const getTestAssets = storage => [
 ];
 
 const addWebStores = storage => {
-    // these `asset => ...` callbacks generate values specifically for the cross-fetch mock
+    // these `asset => ...` callbacks generate values specifically for the fetch mock
     // in the real world they would generate proper URIs
     storage.addWebStore(
         [storage.AssetType.Project],
